pyproject.toml
[project]
name = "ocr-pdf-agent"
version = "0.1.0"
description = "OCR PDF Agent: PDF->JSON/Markdown + Chroma RAG Memory + QA CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "chromadb>=0.5.0",
  "fastapi>=0.115.0",
  "numpy>=1.26",
  "openai>=1.0.0",
  "pillow>=10.0.0",
  "pydantic>=2.7.0",
  "pymupdf>=1.24.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
ocr = [
  "paddleocr>=2.7.0",
]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
ocr = "ocr_agent.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
packages = ["ocr_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
